+++ foxes/models/turbine_models/__init__.py
@@ -4,3 +4,4 @@ from .set_farm_vars import SetFarmVars
 from .thrust2ct import Thrust2Ct
 from .power_mask import PowerMask
 from .sector_management import SectorManagement
+from .table_factors import TableFactors